Major reworking of how we display flights on the map
we no longer use layers and sources created in mapbox studio.
instead, we are now dynamically setting the data for a layer
when we receive flight data from the api endpoint.
this allows us to set data-driven styles for the flight dots.
Flight dots now have a color corresponding to the price.
The map is also much more responsive now